Bugfix, now the esp node MAC is transmitted correctly

+++ greenhouseStar/main/TelemetryPacket.h
@@ -2,7 +2,7 @@
 
 typedef struct {
     uint32_t timestamp;
-    uint8_t node_id;
+    uint32_t node_id;
     float water_temp;
     float tds_value;
     float soil_moisture;
